INTERIM COMMIT - Will compile and build, but is not at all finished.
LightShow is still unusable; the BDF will compile and display properly
but the self-selection mechanism is still based on the clock rather than
an actual routine-dependent signal, as this mechanism is currently not
implemented. Its detritus will need cleaned, as there are still remnants
of my attempt in the Routine* modules.

Furthermore, the PRNG is yielding non-random selections. This is most
apparent in rapid-switching states (such as currently) and will need
some form of masking or increased randomization. Perhaps drawing from
non-consecutive bits will break up the result pattern?